feat: 딜러의 수익을 확인하는 기능 구현
