+++ README.md
@@ -104,7 +104,7 @@ graph TD
 
 - [x] 플레이어가 블랙잭 보너스를 획득한다.
 - [x] 플레이어가 베팅에 실패한다.
-- [ ] 딜러의 수익을 확인한다.
+- [x] 딜러의 수익을 확인한다.
 - [x] 플레이어의 수익을 확인한다.
 
 ### 입력
